GBLD-1035: abort with an error when unequal profile activations should be merged
After merging profiles with the same id, only one profile's activation remains for
the merged profile. The other activation is discarded!

This is probably because it is not sensible to merge profiles with the same id,
but different activations. Example:
profile A is activated by os arch win32, and another profile with same id A is activated
by os arch linux.

Or one is activated by property foo and another one by property bar.

So when we are asked to merge two profiles with the same id but different activations,
we will abort with an error to avoid hard to detect build problems (e.g. profile not
getting activated at all).

Note: only the implementation of the anonymous DefaultInheritanceAssembler.InheritanceModelMerger
is affected, the rest are automatic code reformattings.